Fix broken rowData contains all zero check
The reduce() method runs a function on each array element, storing the return value of each iteration in the accumulator, and finally returns the accumulator value. If you never reference the accumulator value inside the function, then each iteration overwrites the previous result. The test as written was only actually checking whether the final dataRow value was zero.

Refactored to use the every() method which was designed for this exact use case.